+++ src/main/java/com/mrporter/pomangam/test/data/point/PointRankData.java
@@ -13,7 +13,7 @@ public class PointRankData {
     PointRankJpaRepository pointRankJpaRepository;
 
     @Transactional
-    public void of(Long idx, String title, int level, int percentSavePoint, int priceSavePoint, int rewardCouponPrice, int nextLowerLimitOrderCount, int nextLowerLimitRecommendCount) {
+    public void of(Long idx, String title, int level, float percentSavePoint, int priceSavePoint, int rewardCouponPrice, int nextLowerLimitOrderCount, int nextLowerLimitRecommendCount) {
         PointRank pointRank = PointRank.builder()
                 .idx(idx)
                 .level((short) level)
